[Flink] Combine before GBK
